+++ trace-forward/src/Trace/Forward/Run/TraceObject/Forwarder.hs
@@ -11,6 +11,7 @@ import           Ouroboros.Network.Mux (MiniProtocolCb (..), RunMiniProtocol (..
 import           Ouroboros.Network.Util.ShowProxy (ShowProxy (..))
 
 import qualified Codec.Serialise as CBOR
+import           Control.DeepSeq (NFData)
 import qualified Data.ByteString.Lazy as LBS
 import           Data.Void (Void)
 
@@ -18,11 +19,12 @@ import           Trace.Forward.Configuration.TraceObject (ForwarderConfiguration
 import qualified Trace.Forward.Protocol.TraceObject.Codec as Forwarder
 import qualified Trace.Forward.Protocol.TraceObject.Forwarder as Forwarder
 import           Trace.Forward.Utils.TraceObject
-import Trace.Forward.Utils.ForwardSink (ForwardSink)
+import           Trace.Forward.Utils.ForwardSink (ForwardSink)
 
 forwardTraceObjectsInit
   :: (CBOR.Serialise lo,
-      ShowProxy lo)
+      ShowProxy lo,
+      NFData lo)
   => ForwarderConfiguration lo
   -> ForwardSink lo
   -> RunMiniProtocol 'Mux.InitiatorMode initiatorCtx responderCtx LBS.ByteString IO () Void
@@ -31,7 +33,8 @@ forwardTraceObjectsInit config sink =
 
 forwardTraceObjectsResp
   :: (CBOR.Serialise lo,
-      ShowProxy lo)
+      ShowProxy lo,
+      NFData lo)
   => ForwarderConfiguration lo
   -> ForwardSink lo
   -> RunMiniProtocol 'Mux.ResponderMode initiatorCtx responderCtx LBS.ByteString IO Void ()
@@ -39,7 +42,7 @@ forwardTraceObjectsResp config sink =
   ResponderProtocolOnly $ runPeerWithSink config sink
 
 runPeerWithSink
-  :: (ShowProxy lo, CBOR.Serialise lo)
+  :: (ShowProxy lo, CBOR.Serialise lo, NFData lo)
   => ForwarderConfiguration lo
   -> ForwardSink lo
   -> MiniProtocolCb ctx LBS.ByteString IO ()